+++ packages/flutter_blue_ultra/example/lib/screens/bluetooth_off_screen.dart
@@ -23,7 +23,7 @@ class BluetoothOffScreen extends StatelessWidget {
     String? state = adapterState?.toString().split(".").last;
     return Text(
       'Bluetooth Adapter is ${state ?? 'not available'}',
-      style: Theme.of(context).primaryTextTheme.titleSmall?.copyWith(color: Colors.white),
+      style: Theme.of(context).textTheme.titleSmall?.copyWith(color: Colors.white),
     );
   }